create ui notification screen
