+++ app/src/main/res/layout/item_notification.xml
@@ -1,0 +1,127 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+        <variable
+            name="match"
+            type="com.sunasterisk.loltournaments.data.model.remote.Match" />
+
+        <import type="com.sunasterisk.loltournaments.R" />
+    </data>
+
+    <androidx.cardview.widget.CardView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="@dimen/dp_5"
+        app:cardCornerRadius="@dimen/dp_10">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:padding="@dimen/dp_10">
+
+            <TextView
+                android:id="@+id/textMatchTitleName"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:text="@{match.getTitleNotification()}"
+                android:textAlignment="center"
+                android:textColor="@color/color_river_bed"
+                android:textSize="@dimen/sp_20"
+                android:textStyle="bold"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                tools:text="Finals:" />
+
+            <TextView
+                android:id="@+id/textScore"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@{match.getScore()}"
+                android:textColor="@color/color_river_bed"
+                android:textSize="@dimen/sp_25"
+                android:textStyle="bold"
+                android:layout_marginTop="@dimen/dp_10"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                tools:text="VS" />
+
+            <ImageView
+                android:id="@+id/imageFirstTeam"
+                loadImageUrl="@{match.getFirstTeam().imageUrl}"
+                android:layout_width="@dimen/dp_55"
+                android:layout_height="0dp"
+                android:layout_marginEnd="@dimen/dp_10"
+                app:layout_constraintBottom_toBottomOf="@id/textScore"
+                app:layout_constraintDimensionRatio="1:1"
+                app:layout_constraintEnd_toStartOf="@+id/textScore"
+                app:layout_constraintTop_toTopOf="@+id/textScore"
+                tools:src="@tools:sample/avatars" />
+
+            <TextView
+                android:id="@+id/textFirstTeamAcronym"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="@dimen/dp_10"
+                android:text="@{match.getFirstTeam().acronym}"
+                android:textAlignment="center"
+                android:textColor="@color/color_river_bed"
+                android:textSize="@dimen/sp_30"
+                android:textStyle="bold"
+                app:layout_constraintBottom_toBottomOf="@+id/textScore"
+                app:layout_constraintEnd_toStartOf="@+id/imageFirstTeam"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="@+id/textScore"
+                tools:text="DK" />
+
+            <ImageView
+                android:id="@+id/imageSecondTeam"
+                loadImageUrl="@{match.getSecondTeam().imageUrl}"
+                android:layout_width="@dimen/dp_55"
+                android:layout_height="0dp"
+                android:layout_marginStart="@dimen/dp_10"
+                app:layout_constraintBottom_toBottomOf="@id/textScore"
+                app:layout_constraintDimensionRatio="1:1"
+                app:layout_constraintStart_toEndOf="@id/textScore"
+                app:layout_constraintTop_toTopOf="@+id/textScore"
+                tools:src="@tools:sample/avatars" />
+
+            <TextView
+                android:id="@+id/textSecondTeamAcronym"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginHorizontal="@dimen/dp_10"
+                android:text="@{match.getSecondTeam().acronym}"
+                android:textAlignment="center"
+                android:textColor="@color/color_river_bed"
+                android:textSize="@dimen/sp_30"
+                android:textStyle="bold"
+                app:layout_constraintBottom_toBottomOf="@+id/textScore"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toEndOf="@+id/imageSecondTeam"
+                app:layout_constraintTop_toTopOf="@+id/textScore"
+                tools:text="GEN" />
+
+            <TextView
+                android:id="@+id/textMatchTime"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="@dimen/dp_16"
+                android:textColor="@color/color_river_bed"
+                android:textDate="@{match.scheduledAt}"
+                android:textSize="@dimen/sp_16"
+                android:withHour="@{true}"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/textScore"
+                tools:text="08:00 10-04-2021" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.cardview.widget.CardView>
+</layout>
