+++ app/src/main/res/navigation/nav_graph.xml
@@ -13,6 +13,9 @@
             app:destination="@id/serieDetailsFragment"
             app:enterAnim="@anim/anim_push_in"
             app:popExitAnim="@anim/anim_push_out" />
+        <action
+            android:id="@+id/action_homeFragment_to_notificationFragment"
+            app:destination="@id/notificationFragment" />
     </fragment>
     <fragment
         android:id="@+id/teamsFragment"
@@ -35,7 +38,11 @@
     <fragment
         android:id="@+id/moreFragment"
         android:name="com.sunasterisk.loltournaments.ui.more.MoreFragment"
-        android:label="MoreFragment" />
+        android:label="MoreFragment">
+        <action
+            android:id="@+id/action_moreFragment_to_notificationFragment"
+            app:destination="@id/notificationFragment" />
+    </fragment>
     <fragment
         android:id="@+id/serieDetailsFragment"
         android:name="com.sunasterisk.loltournaments.ui.serie.SerieDetailsFragment"
@@ -83,4 +90,8 @@
             android:name="player"
             app:argType="com.sunasterisk.loltournaments.data.model.remote.Player" />
     </fragment>
+    <fragment
+        android:id="@+id/notificationFragment"
+        android:name="com.sunasterisk.loltournaments.ui.notification.NotificationFragment"
+        android:label="NotificationFragment" />
 </navigation>
