+++ app/src/main/res/values/strings.xml
@@ -20,6 +20,7 @@
 
     <string name="text_empty_series">Empty series.</string>
     <string name="text_empty_teams">Empty teams.</string>
+    <string name="text_empty_notification">Empty Notification.</string>
     <string name="text_id">Id:</string>
     <string name="text_start">Start:</string>
     <string name="text_end">End:</string>
